requirements.txt
requests
json
sseclient-py == 1.7.2